requirements.txt
absl-py==1.4.0
affine==2.4.0
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bqplot==0.12.42
branca==0.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colour==0.1.5
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dask==2024.5.2
dask-geopandas==0.3.1
debugpy==1.8.0
decorator==5.1.1
distributed==2024.5.2
dm-tree==0.1.8
earthengine-api==0.1.386
eerepr==0.0.4
etils==1.6.0
exceptiongroup==1.2.0
executing==2.0.1
fiona==1.9.5
flatbuffers==23.5.26
folium==0.15.1
fonttools==4.47.2
fsspec==2023.12.2
future==0.18.3
gast==0.5.4
geemap==0.30.4
geocoder==1.38.1
geopandas==0.14.2
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h5py==3.10.0
httplib2==0.22.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.0
ipyleaflet==0.18.1
ipython==8.20.0
ipytree==0.2.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
kagglehub==0.1.6
keras==2.15.0
keras-core==0.1.7
keras-cv==0.8.1
kiwisolver==1.4.5
libclang==16.0.6
locket==1.0.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.3
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.2
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyogrio==0.7.2
pyparsing==3.1.1
pyperclip==1.8.2
pyproj==3.6.1
pyshp==2.3.1
python-box==7.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.9
ratelim==0.1.6
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
scooby==0.9.2
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-datasets==4.9.4
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
traittypes==0.2.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xyzservices==2023.10.1
zict==3.0.0
zipp==3.17.0
